fix gc, add DD_PBL_IOS define

+++ dd_publishing.h
@@ -100,7 +100,7 @@ bool dd_pbl_ios_reachability_is_wan_aviable();
 void dd_pbl_ios_localytics_session_start(const char * key);
 void dd_pbl_ios_localytics_session_end();
 void dd_pbl_ios_localytics_session_resume();
-void dd_pbl_ios_localytics_track_event(const char * name, size_t attributes_count, const char ** parameters, const char ** values, uint32_t customer_value_increase);
+void dd_pbl_ios_localytics_track_event(const char * name, uint16_t attributes_count, const char ** parameters, const char ** values, uint32_t customer_value_increase);
 void dd_pbl_ios_localytics_track_screen(const char * name);
 
 // -------------------------------------- TestFlight
